+++ CHANGELOG.md
@@ -1,7 +1,14 @@
 # Changelog
 
+## [Unreleased]
+### Added
+- DOMParser for Content Scripts
+
+### Fixed
+- Youtube transcript extraction
+
 ## [0.22.1] - 2026-03-06
-### Addded
+### Added
 - Pinboard import adapter, by [@jlecour](https://codeberg.org/jlecour)
 
 ### Improved